refactor(device): replace AlCtx Deref with explicit base field; sweep dead surface, stale docs, and warnings
`AlCtx` no longer implements `Deref<Target = ServiceCtx>` — contexts
are not smart pointers (C-DEREF), and the implicit fall-through made
it unclear which half of the context a handler touches. The inner
context is now a public `base` field; the ~112 lean accesses in the
AL services spell `ctx.base.…` explicitly. The unused `service_ctx()`
accessor is gone with it.

`CemiTransportLayer` now captures its own `&LayerContext` at
construction instead of reaching through the inner `TransportLayer`,
so the TL's two `pub(crate)` context accessors are deleted; the TL's
own allocation sites use the `self.lctx.buffer_manager` field
directly like every other layer.

`KnxNetIpSecureContext` and its supertraits stay (they back the
upcoming KNXnet/IP Secure dispatch) but carry `#[allow(dead_code)]`
until those arms land.

Warning sweep — the workspace now builds `--all-targets` with only
the two deliberate warnings (`generic_const_exprs`,
`type_alias_bounds`):

- The no-op logging facade discarded its arguments, so every binding
  used only in a log statement warned when no backend feature was
  enabled. The noop macro now references its arguments (free at
  runtime), fixing all such sites at once. Level re-exports carry
  `#[allow(unused_imports)]` for facade completeness across feature
  sets.
- `knx_stack_config!` ends each generated table build with
  `assert!(idx == Self::…_SIZE)`. This consumes the final loop
  increment (previously an unused-assignment warning at every macro
  use) and doubles as a compile-time fill-exactness check, since
  `new()` runs in const contexts.
- `test_basic_config` exercises `create_tables` against the data
  accessors — real coverage instead of an `#[allow(dead_code)]`.
- Fixed the two broken `zweidraehte-client` examples (`u8` → `u16`
  property id, from the extended-property wire-format change) and
  dropped their SESSION.md entry.

Docs brought back in line with the code: removed leftover
`poll_augments` / `next_augment_deadline` listings, the
`LayerContext`-provides-`BufferManagerContext` claims, wrong pub/sub
subscriber counts, "deferred outbox" references, and stale
`LayerStack` / `InsecureDeviceLayers` names; tombstone comments
narrating removed designs were trimmed to present-tense rationale.

Verified: conformance 511/511, device/macros tests, all-targets
workspace build, cross/picow build.

+++ crates/zweidraehte-client/examples/function_property.rs
@@ -90,7 +90,7 @@ struct Config {
     server_addr: SocketAddrV4,
     device_addr: IndividualAddress,
     object_idx: u8,
-    property_id: u8,
+    property_id: u16,
     service_data: Vec<u8>,
 }
 
@@ -98,7 +98,7 @@ fn parse_args(args: &[String]) -> Result<Config, Box<dyn std::error::Error>> {
     let mut server_addr: Option<SocketAddrV4> = None;
     let mut device_addr: Option<IndividualAddress> = None;
     let mut object_idx: Option<u8> = None;
-    let mut property_id: Option<u8> = None;
+    let mut property_id: Option<u16> = None;
     let mut service_data = Vec::new();
 
     let mut i = 1;
